# server/swagger/index.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rentify API",
    "version": "1.0.0",
    "description": "API for Rental Platform"
  },
...
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/blog/auth": {
      "get": {
        "tags": ["Blogs"],
        "summary": "Get authorized blogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/rentPost/auth": {
      "get": {
        "tags": ["RentPosts"],
        "summary": "Get Rent Post (Authorized)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}